fix(geo): change timeZoneOffset to double in ActitoRegion

+++ actito_geo/lib/src/models/actito_region.dart
@@ -13,7 +13,7 @@ class ActitoRegion {
   final int? major;
   final double distance;
   final String timeZone;
-  final int timeZoneOffset;
+  final double timeZoneOffset;
 
   ActitoRegion({
     required this.id,
